Modify requests to use an InputStream for the body
HttpURLConnection uses an OutputStream for writing to the request body.
This introduces some confusion because it isn't clear when the request
is actually sent.

So instead, BoxAPIRequest allows for an InputStream (or String) to be
set which will later be piped to the connection's OutputStream when
send() is called. This makes BoxAPIRequest's interface a lot clearer,
since we can guarantee that network activity will only happen after
send() is called.