Implement exponential backoff and auto-retry
Requests will be automatically retried if the server returns a 500+
response code or a 429. Exponential backoff is also used to wait between
retries.

Auto-retry can be disabled by setting the max number of request attempts
to 1 in BoxAPIConnection.

	api.setMaxAttempts(1);

If the max number of request attempts isn't set, it will default to 3.

Closes #8.