Allow custom fields when getting child items
Iterating over a BoxFolder returns the default fields for each child. In
cases where you want all the information about children, it isn't
practical to call `getInfo()` for each child.

To solve this, two `getChildren()` and `getChildren(String...)` methods
were introduced. The first method returns an iterable that is identical
to iterating over the folder directly. The second method returns an
iterable that retrieves the specified fields for each child item.

Closes #17.