Add shared item support
Add some additional fields to BoxAPIConnection for a shared link and a
shared link password. When a shared item is retrieved, the API
connection will be "forked" into a new API connection that authenticates
using the shared link URL. This way, all subsequent method calls on the
shared item will also authenticate using the shared link.

Closes #40.