Improvements to connection pooling
There were a few bugs identified that were preventing HTTP connections
from being kept alive.

The first bug is the `close()` method in `ProgressInputStream` and
`ProgressOutputStream` weren't closing the underlying stream. This is
now fixed and both classes will now correctly close the underlying
stream.

Anoter issue is that Java won't reuse a connection if the response is
chunked and not read. This means that the response stream can't just be
closed without reading to the end of it. We now ensure that the entirety
of a response's body is read when `disconnect()` is called.

Similarly, there's a bug when wrapping the response's `InputStream` in a
`GZIPInputStream` where the `GZIPInputStream` doesn't read all the way
to the end of a chunked response. Again, the fix for this is to ensure
that the entirety of the raw response stream is read when `disconnect()`
is called.

Java has another bug where automatically handled redirects don't keep
the connection alive. The workaround is to disable automatic redirects
and handle them ourselves by manually creating a new connection.