Add EventLog for retrieving a static log of events
The EventLog class provides methods for retrieving static logs of
events. Unlike the EventStream, an EventLog isn't updated in real-time.
For now it only provides a way of retrieving enterprise events, since
those aren't available in real-time via EventStream.

It's also worth noting that there's a bug in the Box events endpoint
where some enterprise event sources don't map correctly to a
BoxResource. In the cases where this happens, the `getSourceInfo()` will
return null, and a new a `getSourceJSON()` method can be used to access
the event source JSON directly.

Closes #81.