Improve saving and restoring BoxAPIConnections
This change adds a `save()` method and a static `restore(String)` method
to BoxAPIConnection. These methods are meant to simplify the process of
saving the state of a connection and then restoring it at a later time.

The `save()` method serializes the state of BoxAPIConnection to a JSON
string which can persisted. The `restore(String)` method takes a JSON
string and reconstructs a BoxAPIConnection using the saved state.

BoxAPIConnectionListener has also had an
`onError(BoxAPIConnection, BoxAPIException)` callback added so that
connection errors can be more easily handled. The `onRefresh()` method
also now receives a BoxAPIConnection object that can be used to save the
newly updated state of the connection.

Closes #104 and fixes #82.