Handle groups when retrieving and creating collabs
Since the "group" type wasn't being handled when creating a new
collaboration, an IllegalArgumentException would be thrown. This fix
handles group types correctly and will successfully create group-based
collaborations.

When retreiving collaborations, the JSON was parsed as a list of users
instead of a list of users or groups. This led to BoxUser.Info objects
being returned with incorrect or unpopulated fields. This fix makes it
so the "type" field is checked for each item in the list, and the JSON
gets parsed into its appropriate type.

Fixes #116.