Return all results for BoxGroup.getMemberships()
This method wasn't correctly paginating results and only returned the
first 100 memberships. This change returns all of them by automatically
paging.

Ideally, we'd like to return an Iterator that allows for lazy paging,
but that would require a method signature change. This is tracked as a
breaking change in #335.

This change was originally authored by Dor Fire, but modified to prevent
the backwards-incompatible change.

Fixes #316.