Quick fix for NPE on request.toString() in interceptor
Small fix to prevent BoxAPIRequest#toString() from throwing a null
pointer exception inside of a request interceptor.  Ideally, the
request headers would always be available on the request object
instead of being partially stored on the connection, but in lieu
of that more involved fix, this should restore at least some
functionality in the meantime.