Ensure that chunked upload fills buffer from stream
When reading bytes from the input stream for chunked upload, the
SDK was only calling stream.read() once per chunk.  However, the
stream is not guaranteed to provide exactly the number of bytes
requested.  The SDK should track how many bytes have been read
and call read() in a loop to ensure the correct number of bytes
are extracted from the stream.